rename `pkg/sandboxes` -> `pkg/sandbox`
